Refactor BaseChildImmunization class :recycle: - Refactor for testability and maintainability
