+++ build.gradle
@@ -38,7 +38,7 @@ subprojects {
     ext.androidBuildToolsVersion = '29.0.2'
     ext.androidMinSdkVersion = 18
     ext.androidCompileSdkVersion = 28
-    ext.androidTargetSdkVersion = 28
+    ext.androidTargetSdkVersion = 29
 
     ext.androidAnnotationsVersion = '3.0.1'
     ext.androidAnnotationsAPIVersion = '3.0.1'
